requirements.txt
django
djangorestframework
pillow
django-cors-headers

asgiref
django-environ
djangorestframework-simplejwt
PyJWT
pytz
sqlparser
tzdata